Trained a model
I trained a model by yoinking code and making none of it myself. It was this guy's:
https://www.kaggle.com/prabhavsingh/urbansound8k-classification

But now I have a trained model which predicts sounds accurately with 86% accuracy, will keep working on this tomorrow.

If you want to test it yourself you need the data set from here (7gb): https://www.kaggle.com/chrisfilo/urbansound8k

You need to put the script in the same directory as all the folders (also extract all the folders.